+++ pkg/distributor/query.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"sort"
 	"time"
 
 	"github.com/opentracing/opentracing-go"
@@ -52,6 +53,33 @@ func (d *Distributor) Query(ctx context.Context, from, to model.Time, matchers .
 		return nil
 	})
 	return matrix, err
+}
+
+func (d *Distributor) QueryExemplars(ctx context.Context, from, to model.Time, matchers ...[]*labels.Matcher) (*ingester_client.ExemplarQueryResponse, error) {
+	var result *ingester_client.ExemplarQueryResponse
+	err := instrument.CollectedRequest(ctx, "Distributor.QueryExemplars", d.queryDuration, instrument.ErrorCode, func(ctx context.Context) error {
+		req, err := ingester_client.ToExemplarQueryRequest(from, to, matchers...)
+		if err != nil {
+			return err
+		}
+
+		// We ask for all ingesters without passing matchers because exemplar queries take in an array of array of label matchers.
+		replicationSet, err := d.GetIngestersForQuery(ctx, nil)
+		if err != nil {
+			return err
+		}
+
+		result, err = d.queryIngestersExemplars(ctx, replicationSet, req)
+		if err != nil {
+			return err
+		}
+
+		if s := opentracing.SpanFromContext(ctx); s != nil {
+			s.LogKV("series", len(result.Timeseries))
+		}
+		return nil
+	})
+	return result, err
 }
 
 // QueryStream multiple ingesters via the streaming interface and returns big ol' set of chunks.
@@ -106,7 +134,7 @@ func (d *Distributor) GetIngestersForQuery(ctx context.Context, matchers ...*lab
 	}
 
 	// If "shard by all labels" is disabled, we can get ingesters by metricName if exists.
-	if !d.cfg.ShardByAllLabels {
+	if !d.cfg.ShardByAllLabels && len(matchers) > 0 {
 		metricNameMatcher, _, ok := extract.MetricNameMatcherFromMatchers(matchers)
 
 		if ok && metricNameMatcher.Type == labels.MatchEqual {
@@ -183,6 +211,82 @@ func (d *Distributor) queryIngesters(ctx context.Context, replicationSet ring.Re
 	}
 
 	return result, nil
+}
+
+// mergeExemplarSets merges and dedupes two sets of already sorted exemplar pairs.
+// Both a and b should be lists of exemplars from the same series.
+// Defined here instead of pkg/util to avoid a import cycle.
+func mergeExemplarSets(a, b []cortexpb.Exemplar) []cortexpb.Exemplar {
+	result := make([]cortexpb.Exemplar, 0, len(a)+len(b))
+	i, j := 0, 0
+	for i < len(a) && j < len(b) {
+		if a[i].TimestampMs < b[j].TimestampMs {
+			result = append(result, a[i])
+			i++
+		} else if a[i].TimestampMs > b[j].TimestampMs {
+			result = append(result, b[j])
+			j++
+		} else {
+			result = append(result, a[i])
+			i++
+			j++
+		}
+	}
+	// Add the rest of a or b. One of them is empty now.
+	result = append(result, a[i:]...)
+	result = append(result, b[j:]...)
+	return result
+}
+
+// queryIngestersExemplars queries the ingesters for exemplars.
+func (d *Distributor) queryIngestersExemplars(ctx context.Context, replicationSet ring.ReplicationSet, req *ingester_client.ExemplarQueryRequest) (*ingester_client.ExemplarQueryResponse, error) {
+	// Fetch exemplars from multiple ingesters in parallel, using the replicationSet
+	// to deal with consistency.
+	results, err := replicationSet.Do(ctx, d.cfg.ExtraQueryDelay, func(ctx context.Context, ing *ring.InstanceDesc) (interface{}, error) {
+		client, err := d.ingesterPool.GetClientFor(ing.Addr)
+		if err != nil {
+			return nil, err
+		}
+
+		resp, err := client.(ingester_client.IngesterClient).QueryExemplars(ctx, req)
+		d.ingesterQueries.WithLabelValues(ing.Addr).Inc()
+		if err != nil {
+			d.ingesterQueryFailures.WithLabelValues(ing.Addr).Inc()
+			return nil, err
+		}
+
+		return resp, nil
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	// Merge results from replication set.
+	var keys []string
+	exemplarResults := make(map[string]cortexpb.TimeSeries)
+	for _, result := range results {
+		r := result.(*ingester_client.ExemplarQueryResponse)
+		for _, ts := range r.Timeseries {
+			lbls := cortexpb.FromLabelAdaptersToLabels(ts.Labels).String()
+			e, ok := exemplarResults[lbls]
+			if !ok {
+				exemplarResults[lbls] = ts
+				keys = append(keys, lbls)
+			}
+			// Merge in any missing values from another ingesters exemplars for this series.
+			e.Exemplars = mergeExemplarSets(e.Exemplars, ts.Exemplars)
+		}
+	}
+
+	// Query results from each ingester were sorted, but are not necessarily still sorted after merging.
+	sort.Strings(keys)
+
+	result := make([]cortexpb.TimeSeries, len(exemplarResults))
+	for i, k := range keys {
+		result[i] = exemplarResults[k]
+	}
+
+	return &ingester_client.ExemplarQueryResponse{Timeseries: result}, nil
 }
 
 // queryIngesterStream queries the ingesters using the new streaming API.